// app/src/api/streak_board/StreakBoardClient.ts
export interface Habit {
  id: string;
  author: string;
  title: string;
  current_streak: number;
  longest_streak: number;
  last_check_in_date: string;
  created_at: number;
}